Active Directory sAMAccountName limit is 20 characters
AD's limit is 20 characters, including trailing '$' for computer
objects. The length check code is testing the entire sAMAccountName
value, including the '$' character. Bump MAX_SAM_ACCOUNT_LEN to match
this.

This change allows msktutil to create computer objects with names 19
characters in length.

Thanks to Jaroslaw Polok for the bug report and the fix.

FIXES 13