conditionalize the ability to disable reverse lookups for LDAP
Not all OpenLDAP libraries support LDAP_OPT_X_SASL_NOCANON. The
LDAP_OPT_X_SASL_NOCANON flag was added to OpenLDAP in Nov 2008. In
particular, RHEL 5's OpenLDAP does not support it.

Add an ifdef conditional so we can still build on RHEL 5, and display a
message to the user if the LDAP library is too old.