Safely use ns_initparse by linking statically
Linking libresolv statically for ns_initparse and having it pulled in
dynamically by some other lib is not safe and asks for trouble. Instead
detect that we need it linked statically and advise the user to link the
whole program statically. This still is less then ideal but at least
safe.