implement new workflow for keytab updates
when doing password changes, msktutil is currently waiting for these
changes to be reflected in AD. In some AD environments this may take
lots of time due to very slow replication.

The current behaviour is necessary because the msds-KeyVersionNumber
attribute may also take a lot of time to be updated and without
sleeps ldap.get_kvno() would return the old value. However this introduces
a race condition: while msktutil is performing its sleep() loops the new
keys are not present in the keytab. Tickets that are issued in this time
cannot be validated by the keytab. Therefore a different approach is needed.

This patch introduces a different behaviour: it calls ldap.get_kvno()
before changing the password and simply increments the returned
value. All sleeps are removed.