update_keytab(): really add principal with short hostname
Initial commit 1d6a29973eb400979ac83754ede51e1796079f29 used
host/<sAMAccountName> which commonly, but not always matches the short
hostname. This changes brings the behaviour in line with what was originally
intended, and avoids adding a potentially invalid (ie. unknown in AD) SPN if
the short hostname doesn't match the sAMAccountName.

Ticket validation in sssd is the one prominent example that's prone to break if
such invalid SPNs are present in /etc/krb5.keytab.