Check ber_pvt_opt_on in lber and fail if not found
Configure script does not fail on HP-UX, but simply warns if lber isn't found.
This might be overlooked and result in a link failure. Rather fail if not
available because we need LBER_OPT_ON. The workaround for HP-UX is not necessary
anymore, it simply works as expected on HP-UX 11.31.