msktname: improve robustness to DC errors
Improve DC validation to check LDAP connectivity even if we're told to
not canonify hostnames. Skip DCs that do not pass validation checks,
and move on to the next instead of terminating with an error.
In order to do this, DNS handling was refactored into two new classes,
and (mostly) pushed into a new header file msktname.h.