dns_lookup_kdc setting in create_fake_krb5_conf causes issue when using a trusted domain user to create computer account (#208)
When an account outside of the default realm is used Kerberos is not able to
follow a cross-realm referral since the intermediate or target realm do not have
a configuration block in the `[realms]` section. When setting `dns_lookup_kdc`
to `true` Kerberos will still use our fixed (manually discovered) KDC for our
default realm, but will resort to DNS for any other realm to follow referrals.

This fixes #208