configure.ac: fix linking with minimal LIBS
We directly call function `error_message()` from
`libcom_err`, and therefore have to explicitly link
the msktutil binary to it. This used to work because
`krb5-config --libs` always included the necessary
`-lcom_err`, but for dynamic linking, recent versions
started to omit transitive dependencies, and merely
return the minimal `-lkrb5` instead.

The explicit check for `libcom_err` was removed in
189f8b356ab78936cbfad2b1d4d827cdc0028341 as part of
Solaris-related fixes. However, this part of the
commit is dubious because the check itself looks
correct and required. It is unclear why Solaris builds
would break, but in any case, a different fix is
appropriate here.

This change reverts a part of commit 189f8b356ab78936cbfad2b1d4d827cdc0028341.
It addresses a build problem discussed as part of issue #211
(but is not related to the main topic of this issue).

Signed-off-by: Daniel Kobras <kobras@puzzle-itc.de>