configure.ac: more robust krb5 autodetection
In order to distinguish between Heimdal and MIT, we used to
check if some Heimdal-specific headers are present on the
build system. Some distros allow both implementations to be
installed in parallel, though, so this heuristic becomes
fragile. Recent krb5-config supports the --vendor option,
which provides a more reliable check. Fall back to the old
method if no vendor is reported.